refactor: move table metadata alongside resolver
We already have a metadata resolver, so lets stick the metadata types
next to it.
