+++ ingester/src/buffer_tree/partition/resolver/coalesce.rs
@@ -13,7 +13,9 @@ use hashbrown::{hash_map::Entry, HashMap};
 use parking_lot::Mutex;
 
 use crate::{
-    buffer_tree::{namespace::NamespaceName, partition::PartitionData, table::TableMetadata},
+    buffer_tree::{
+        namespace::NamespaceName, partition::PartitionData, table::metadata::TableMetadata,
+    },
     deferred_load::DeferredLoad,
 };
